refactor: move facade interfaces to shared module
