+++ src/php/Interop/Domain/ExportFinder/FunctionsToExportFinder.php
@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace Phel\Interop\Domain\ExportFinder;
 
 use Phel;
-use Phel\Build\BuildFacadeInterface;
 use Phel\Build\Domain\Extractor\ExtractorException;
 use Phel\Build\Domain\Extractor\NamespaceInformation;
-use Phel\Command\CommandFacadeInterface;
 use Phel\Compiler\Domain\Evaluator\Exceptions\CompiledCodeIsMalformedException;
 use Phel\Compiler\Domain\Evaluator\Exceptions\FileException;
 use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Interop\Domain\ReadModel\FunctionToExport;
 use Phel\Lang\Collections\Map\PersistentMapInterface;
 use Phel\Lang\Keyword;
+use Phel\Shared\Facade\BuildFacadeInterface;
+use Phel\Shared\Facade\CommandFacadeInterface;
 
 final readonly class FunctionsToExportFinder implements FunctionsToExportFinderInterface
 {
