+++ src/php/Run/Application/EvalExecutor.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Phel\Run\Application;
 
-use Phel\Compiler\CompilerFacadeInterface;
 use Phel\Compiler\Domain\Evaluator\Exceptions\CompiledCodeIsMalformedException;
 use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Compiler\Domain\Parser\Exceptions\UnfinishedParserException;
@@ -12,6 +11,7 @@ use Phel\Compiler\Infrastructure\CompileOptions;
 use Phel\Printer\PrinterInterface;
 use Phel\Run\Domain\Repl\ColorStyleInterface;
 use Phel\Run\Domain\Repl\ReplCommandIoInterface;
+use Phel\Shared\Facade\CompilerFacadeInterface;
 use Throwable;
 
 use function array_reverse;
