+++ src/php/Run/RunFacade.php
@@ -8,6 +8,7 @@ use Gacela\Framework\AbstractFacade;
 use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Compiler\Infrastructure\CompileOptions;
+use Phel\Shared\Facade\RunFacadeInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
