+++ src/php/Shared/README.md
@@ -1,10 +1,11 @@
 # Shared Module
 
 ## Purpose
-Common constants and utilities shared across all modules.
+Common constants, utilities, and facade interfaces shared across all modules.
 
 ## Responsibilities
 - Defining build-related constants (compile mode, build mode)
 - Compiler constants (core namespace definitions)
 - REPL-specific constants
 - Cross-module shared values and configuration
+- Centralizing all module facade interfaces in `Facade/` directory
